364.2. RTMP Server Setup(Section 26: Implementing Streaming Video)
So I've got a new code editor open here inside of RCMP server.

There's the package on file that we just created and there's the node modules directory where our packages

are currently being installed.

So inside if you're going to create a new file called index dot J.S. I'm going to go back over to the

Get help repository copy this example code right here.

And then paste it inside the index not just file

and then going to save this file and then to start up our project.

I'm going to open up the packaged case on file and I will add in a new script inside of your and replace

the existing test script to start up our project.

So I'm going to delete test and I'm going to add in a new script called start anytime we run that script.

We're going to run node index Dutch.

Yes like the OK.

And that's pretty much it.

So now if I flip back over to my terminal I should see my package is finished installing the node medius

server package.

So inside of our server I'm going to run NPM starts and that's going to start up our streaming server.

So I should see a message like this node media server whatever version you are running.

And then we'll see some other output in here over time.

So now our server for streaming is running on port eight thousand.

OK.

So this looks pretty good.

So let's take a quick pause right here.

When we come back the next section we're going to start to configure our OBSS software to or whatever

streaming software you're using to stream to the server.

So quick pause and I'll see you in just a minute.