368.2. Implementing FLV JS(Section 26: Implementing Streaming Video)
So let's add a couple of additional props to that video element to get it to work a little bit better

than it is right now to do so.

We're going to find the video element again.

I'm going to give it a style prop and I'm going to assign it a whiff of 100 percent.

Now a quick reminder here we have the two sets of curly braces because the outer set means that we're

going to pass in or define a javascript expression the inner set is defining an actual object.

We're then going to give it a width of 100 percent.

And remember you need to put that 100 percent inside of quotes.

I'm going to add on one other prop to this thing as well.

I'll say controls equals.

True like so.