368.3. Implementing FLV JS(Section 26: Implementing Streaming Video)
Now remember any time that we pass a prop down to a child component and it's getting a value of simply

true we don't actually have to write out true.

We can simply write out the name of the prop and that's going to show up as controls equals true inside

of that child.

All right.

Let's now try saving this and then flipping back over.

And now when we do so we can see a much more standard video player.

OK so this looks pretty good.

Let's take another quick pause right here.

When we come back the next section we're going to use the F.L. V library and hook that thing up to the

video player that we just created.

So quick pause and I'll see you in just a minute.