370.1. Optional Player Building(Section 26: Implementing Streaming Video)
An last video we set up a bunch of stuff around the Felty player but then we very quickly got confronted

with a nasty error message here.

It's kind of hard to decipher what's going on with this error message but you'll notice in sight of

this component in Mount print up right here it appears that things are starting to go wrong.

When we tried to attach the media player to the video player ref.

So let's try adding a console log of the video ref inside of component dismount and figure out what's

going wrong.

So inside of component amount going to add a console that log for this video ref and I'll save this

and we'll see what happens.

OK.

So I still see the same error message inside my console over here.

And scroll up just a little bit and I'll see the console log right here.

So it says Current is null.

Now why could that be.

This essentially means that we are not getting the correct reference to that video element.

Well let's kind of think about this a little bit.

If we go down to a render method you'll recall that we have a little bit of a check inside of here and

we say if we do not yet have access to our stream then we're going to just show a div with loading.

So when our component first renders if we don't have the stream we are not creating the video element

at all.

Instead we just immediately return the Stiv right here.

We never create the video and we never attach our ref to that video element.

So that's essentially what's happening when our component first gets rendered.

We're making the div.

The video does not get created because we do not yet have our stream.

However I bet you anything that if we dismiss this message right here or excuse me we don't need to

dismiss it we can just go back to localhost 3000.

And then if we then attempt to click on one of these elements we will see the video player appear successfully.

That's because if we go back to the original list screen we're going to load up the stream with the

idea of one.

And so that's going to be inside of our redux store when we then attempt to render streams show we are

already in control of the appropriate stream.

So we don't encounter this case where we have not yet loaded up.

However if we refresh the page while we are here then we have to wait for the Ajax request or the network

request to be completed to go and get that stream object.

And that's the case where we run into this error message right here.

So how can we fix this.

Well there's more than one way to do it.

We can either say that we do not want to render anything inside this component until we fetch the stream

and essentially continue what we're doing right here and then add in some better logic inside of component

and mount to decide whether or not we are actually ready to attempt to attach the video player to the

video ref.

Alternatively because our render method or the entire component itself only tries to print up the title

and description from the stream.

We could only guard against printing out the H1 and h 5 right here until we actually get the stream.

So in other words always try to show the video player but wait to show the title and description until

we get the stream.

So we get totally do that approach because the video player does not actually depend upon the stream

being fetched.

Now that would be a really easy approach or real easy solution but I want to show you the slightly more

complicated solution where we're going to essentially make sure that we do not attempt to attach that

video player until we successfully get the stream because I think that's more realistic of what's going

to happen in real life.

The reason I say it's more realistic is because when we specify the you are right here raising the ID

which is already known ahead of time before we load up the appropriate stream.

But it's entirely possible that the oil right here might need to reference some property inside of our

stream object in order to side what the actual R L is.

So we might need to reference some like I don't know internal stream or id or streamer tokin or something

like that when specifying the You Are All Right here.

And if we wanted to do that we would definitely need to have access to the stream.

All right.