370.2. Optional Player Building(Section 26: Implementing Streaming Video)
Now that we understand what's going on here I'm going to remove the console lock.

So now inside of this component we only want to attempt to set up this video player right here after

we know that the appropriate stream has been actually fetched.

So here's what we're going to do.

I'm going to define another helper method inside this component and I'm going to call it build.

Player like so then inside if you're going to put a IF statement.

I'm going to say if this player exists in the words if we have already set up the player or if this

Propst stream does not exist.

So notice the not right there.

So if we already have a player or if the stream does not yet exist then just return.

Otherwise we want to go to the process of setting up our player.

So I went to copy all this stuff right here for the star player all the way down to this player to load.

I'm going to cut it and put it inside of billed player

Next up I need to make sure that we have access to the property inside of player so I'll just copy that

destructuring call from right here inside of component and mount down into build player like so.

So now we've got the idea right here.

Easily referenced.

Next up any anytime or component is first mounted to the screen.

We're going to attempt to build the player.

So inside of component in mount I would call this build player.

So now whenever this component is first rendered we're going to attempt to fetch our stream.

We're also going to try to build the player.

So we're going to call build a player if we have not built the player before.

And if we do not have the stream we're going to return.

So that's going to take care of the case of us first loading the page and not yet having access to the

actual Stream.

Now remember component mount is only going to be called one time when the company is first rendered

to the screen.

So just in case we do not yet have access to our stream and we're going to try to fetch it on the fly

we need to have some place to call build player after we successfully fetch the stream.

So to handle that case will define another lifecycle method component.

Did update and inside of here we will also call Bills player.

So now the idea is that when our component first renders attempt to build the player and then if our

component fetches the stream successfully at some point in the future and the component renders component

it update will be called.

And we will attempt to call build player in there as well.

So essentially at any point at this component gets rendered either the initial render or any follow

up render we're always going to attempt to build a player.

But if we've already built it and we already have a player or if we do not have the stream then we're

not going to try to build it.

All right so let's save this and try it out inside of our browser.

So I'll flip back over and I do not see any error message whatsoever.

I can refreshed everything looks good.

I can also go back to the stream index I can refresh the page over here.

Navigate over to a stream and everything still looks good as well.

Perfect.

All right so what you see right here is a good example of how we have some task that we want to do inside

of a component.

And we just need it to happen no matter what be it either when the component first gets rendered or

when it gets re rendered.

At some point in the future.

So I think this looks pretty good.

Let's continue in the next video.

Well we're going to start to test out actual streaming video to our application.